fix: rename findabe.ts -> findable.ts and saveble.ts -> saveable.ts
Corrects typos in filenames and updates all internal imports.
README architecture tree was already correct — reverted the mistaken
change from the previous docs commit.
