+++ src/domain/application/repositories/findable.ts
